Moving nonlin_coefs to mocks.py
The function nonlin_coefs() is used in mocks.py, not in test_kgain_cal.
