+++ corgidrp/mocks.py
@@ -1,5 +1,6 @@
 import os
 import numpy as np
+import pandas as pd
 import astropy.io.fits as fits
 from astropy.time import Time
 
@@ -358,6 +359,47 @@ def create_badpixelmap_files(filedir=None, col_bp=None, row_bp=None):
 
     return badpixelmap
 
+def nonlin_coefs(filename,EMgain,order):
+    """
+    Reads TVAC nonlinearity table from location specified by ‘filename’.
+    The column in the table closest to the ‘EMgain’ value is selected and fits
+    a polynomial of order ‘order’. The coefficients of the fit are adjusted so
+    that the polynomial function equals unity at 3000 DN. Outputs array polynomial
+    coefficients, array of DN values from the TVAC table, and an array of the
+    polynomial function values for all the DN values.
+
+    Args:
+      filename (string): file name
+      EMgain (int): em gain value
+      order (int): polynomial order
+
+    Returns:
+      np.array: fit coefficients
+      np.array: DN values
+      np.array: fit values
+    """
+    # filename is the name of the csv text file containing the TVAC nonlin table
+    # EM gain selects the closest column in the table
+    # Load the specified file
+    bigArray = pd.read_csv(filename, header=None).values
+    EMgains = bigArray[0, 1:]
+    DNs = bigArray[1:, 0]
+
+    # Find the closest EM gain available to what was requested
+    iG = (np.abs(EMgains - EMgain)).argmin()
+
+    # Fit the nonlinearity numbers to a polynomial
+    vals = bigArray[1:, iG + 1]
+    coeffs = np.polyfit(DNs, vals, order)
+
+    # shift so that function passes through unity at 3000 DN for these tests
+    fitVals0 = np.polyval(coeffs, DNs)
+    ind = np.where(DNs == 3000)
+    unity_val = fitVals0[ind][0]
+    coeffs[3] = coeffs[3] - (unity_val-1.0)
+    fitVals = np.polyval(coeffs,DNs)
+
+    return coeffs, DNs, fitVals
 
 def nonlin_factor(coeffs,DN):
     """
